backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CrewAI and AI
crewai==0.1.26
langchain==0.1.4
openai==1.10.0
anthropic==0.8.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache & Queue
redis==5.0.1
celery==5.3.6
kombu==5.3.5

# Storage
boto3==1.34.34
minio==7.2.3

# Git Integration
gitpython==3.1.41
pygithub==2.1.1
python-gitlab==4.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
pyyaml==6.0.1
jinja2==3.1.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
